refactor: define top-level semantics
This reshapes many of the shells and non of the core logic. We still
pull new entries to archive, compress them, and store them. We now
define each of these pieces in a (near) top-level trait.

We make compression its own isolated, concurrently executed step.

We add a channel to acknowledge stored messages.