chore: rename to Mouseion
The Mouseion of Alexandria, which arguably included the Library of
Alexandria, was an institution said to have been founded by Ptolemy
I Soter and his son Ptolemy II Philadelphus. Originally, the word
mouseion meant any place that was dedicated to the Muses, often related
to the study of music or poetry, but later associated with sites of
learning such as Plato's Academy and Aristotle's Lyceum.